feat: refactor tree water needs (#963)
* feat: refactor tree water needs

* feat: useIsInVegetationPeriod as function
