+++ src/components/splash/splash.tsx
@@ -13,15 +13,7 @@ import Markdown from "react-markdown";
 import { SplashTreeIcon } from "../icons/splash-tree-icon";
 import { ExternalAnchorLink } from "../anchor-link/external-anchor-link";
 import { useSplashStore } from "./splash-store";
-import { isWithinInterval } from "date-fns";
-
-const currentDate = new Date();
-const currentYear = currentDate.getFullYear();
-
-const isInVegetationPeriod = isWithinInterval(currentDate, {
-	start: new Date(currentYear, 2, 1), // Beginning of March
-	end: new Date(currentYear, 8, 31), // End of October
-});
+import { useIsInVegetationPeriod } from "../../utils/use-is-in-vegetation-period";
 
 interface SectionHeadingProps {
 	title: string;
@@ -48,6 +40,8 @@ const SectionHeading: React.FC<SectionHeadingProps> = ({
 export const Splash: React.FC = () => {
 	const i18n = useI18nStore().i18n();
 	const { hideSplashScreen } = useSplashStore();
+
+	const isInVegetationPeriod = useIsInVegetationPeriod();
 
 	const splashContainer = useRef<HTMLDivElement | null>(null);
