+++ src/components/tree-detail/tree-water-needs/water-progress-circle/water-progress-circle.tsx
@@ -2,6 +2,7 @@ import React, { useMemo } from "react";
 import { TreeAgeClassification, WaterCircleProps } from "../../tree-types";
 import { useI18nStore } from "../../../../i18n/i18n-store";
 import { PartialProgressCircle } from "./partial-progress-circle";
+import { WaterProgressCircleLegend } from "./water-progress-circle-legend";
 
 export const WaterProgressCircle: React.FC<WaterCircleProps> = ({
 	parts,
@@ -9,6 +10,8 @@ export const WaterProgressCircle: React.FC<WaterCircleProps> = ({
 	shouldBeWatered,
 	treeAgeClassification,
 	size,
+	rainSum,
+	wateringSum,
 }) => {
 	const i18n = useI18nStore().i18n();
 	const { formatNumber } = useI18nStore();
@@ -32,14 +35,23 @@ export const WaterProgressCircle: React.FC<WaterCircleProps> = ({
 	}, [treeAgeClassification, needsWaterAmount, shouldBeWatered, i18n]);
 
 	return (
-		<div
-			className={`grid w-[${size}px] grid-cols-1 grid-rows-1`}
-			data-testid="water-progress-circle"
-		>
-			<PartialProgressCircle
-				parts={parts}
-				title={circleTitle}
-				size={size * 0.9}
+		<div className="flex flex-row items-center justify-start gap-4">
+			<div
+				className={`grid w-[${size}px] grid-cols-1 grid-rows-1`}
+				data-testid="water-progress-circle"
+			>
+				<PartialProgressCircle
+					parts={parts}
+					title={circleTitle}
+					size={size * 0.9}
+				/>
+			</div>
+			<WaterProgressCircleLegend
+				needsWaterAmount={needsWaterAmount}
+				shouldBeWatered={shouldBeWatered}
+				treeAgeClassification={treeAgeClassification}
+				rainSum={rainSum}
+				wateringSum={wateringSum}
 			/>
 		</div>
 	);
